src/requirements.txt
mysqlclient
flask_sqlalchemy
Flask
python-dotenv